Don't add chartLabels in the tooltip if they aren't provided.
It' usefull when constructing pie charts without legends
like the following example:

$chart = new \core\chart_pie();
$s = new \core\chart_series("Name", [10,20,30,40]);
$s->set_labels(["Label 1", "Label 2", "Label 3", "Label 4"]);
$chart->add_series($s);

With the current code the legend is not rendered in the chart but
an undefined is showed in the tooltip when hovering the pie
because the chartLabel hasn't been provided.

A Pie chart without legend is useful when rendering more
pie charts one next to the other. If the legend of the pie is different
each other the pies have different sizes.